All Functions needed for Controllers #change and test with float still needed

+++ Trips/Models/PhoneData.cs
@@ -10,6 +10,6 @@ namespace Trips.Models
         public  Vector4 Orientation;
         public  Vector3 Acceleration;
         public  DateTime TimeStamp;
-        public double RotationAmount;
+        public float RotationAmount;
     }
 }
